requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
apscheduler==3.11.0
requests==2.32.5
PySocks==1.7.1
feedparser==6.0.12
PyYAML==6.0.2
jinja2==3.1.6
cryptography==45.0.7
scrapling==0.4.2
curl_cffi>=0.14.0
playwright==1.58.0
browserforge>=1.2.4
Pillow==11.2.1
langchain-core==1.2.16
langgraph==1.0.10
